Make jruby in 1.9 mode work as a normal ruby
This is to make spawning processes work for jruby 1.7+

Note: Process.spawn does not yet support the writers on jruby, so you
will get warnings about :out and :err and you won't get the output.
That should be fixed in future versions. This will get things working
for now though.