Add support for initscript from foreman-export-initscript.
This previously was a gem and I don't see a reason why it shouldn't be
included the main distribution. Due to it being a gem a lot of code
needs to be replicated based on how Foreman was designed. I think it
needs the love and care of a community!