Clean up upstart export to use native upstart features
The old upstart used a big ugly line with `su` `export` `sh` and other evilness. Aside from this, it was also a significant security vulnerability as anyone with access to the box could run `ps` and see all the variables, including things like database credentials.
This change makes the upstart config use the native features of upstart itself.