Allow supervisord environment vars to have "="
Prior to this PR, an environment variable configured for use with Foreman could not contain a `?`, `=` or `&` (common in URLs) if the target configuration was set to SupervisorD. This PR cleans up the output from `Shellwords.escape` in order to be ready for use in a `app*.conf` file. I've updated the tests to cover this particular issue.

Fixes #571
Fixes #519

/cc @inbeom, @ddollar