[spec_helper.rb] Error when Procfile present
With a present and empty Procfile in the root directory of the project,
this has the potential to cause specs in cli_spec.rb to fail in a
non-obvious way, and is not simple to debug the root cause.

Because of how `forked_foreman` is implemented, it doesn't included the
`FakeFs` context to the subprocess, so it is pretty hard to deduce that
it is running outside of that context, and that any generated `Procfile`
(using `write_procfile` for example) will not be visible to the sub
process.

To avoid this confusion and errors that might be caused from this, this
error is in place to warn the user if the file currently exists and
prevents running the specs with it in place.