kill the output watch thread on shutdown
It was left running, and being re-started next time {#start} is called.
Not a big impact in the standard way to use foreman, but was making
debugging tests very confusing - I would get a lot of spurious no
color method on $stdout because of dangling threads trying to display
stuff after the capture_output returned.